+++ Changelog.md
@@ -23,7 +23,7 @@
 - Added task panel to view the progress of running async tasks
 - Run post processing operations asynchronously
 - Added support for the NanoDLP format (.nanodlp)
-- Allow setting/updating default slice config (#5)
+- Allow setting/updating default slice config ([#5](https://github.com/connorslade/mslicer/issues/5))
 
 ## v0.4.0 &mdash; November 25th, 2025